pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pyinstaller = "^6.14.1"

[project]
name = "auto-reservation"
version = "0.1.0"
description = ""
authors = [
    {name = "bocswrd",email = "menu3121@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = ["altgraph (==0.17.4)", "attrs (==25.3.0)", "certifi (==2025.4.26)", "cffi (==1.17.1)", "charset-normalizer (==3.4.2)", "colorama (==0.4.6)", "greenlet (==3.2.2)", "h11 (==0.16.0)", "idna (==3.10)", "iniconfig (==2.1.0)", "outcome (==1.3.0.post0)", "packaging (==25.0)", "pefile (==2023.2.7)", "pluggy (==1.6.0)", "pycparser (==2.22)", "pyee (==13.0.0)", "pygments (==2.19.1)", "pyinstaller-hooks-contrib (==2025.4)", "pysocks (==1.7.1)", "pytest-base-url (==2.1.0)", "pytest-playwright (==0.7.0)", "python-dotenv (==1.1.0)", "python-slugify (==8.0.4)", "pywin32-ctypes (==0.2.3)", "requests (==2.32.3)", "setuptools (==80.4.0)", "sniffio (==1.3.1)", "sortedcontainers (==2.4.0)", "text-unidecode (==1.3)", "trio (==0.30.0)", "trio-websocket (==0.12.2)", "typing-extensions (==4.13.2)", "urllib3 (==2.4.0)", "webdriver-manager (==4.0.2)", "websocket-client (==1.8.0)", "wsproto (==1.2.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
